swagger configuration moved to shared AspNetCore library
