+++ shared/EShopOnAbp.Shared.Hosting.AspNetCore/SwaggerWithAuthConfigurationHelper.cs
@@ -1,9 +1,9 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using System.Collections.Generic;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
-using System.Collections.Generic;
 using Volo.Abp.Modularity;
 
-namespace EShopOnAbp.Shared.Hosting.Gateways
+namespace EShopOnAbp.Shared.Hosting.AspNetCore
 {
     public static class SwaggerWithAuthConfigurationHelper
     {
